requirement.txt
numpy==1.24.4
scikit-learn==1.3.0
transformers==4.33.2
tqdm==4.66.1
PyYAML==6.0.1
matplotlib==3.7.2
pandas==2.0.3
